// src/types/types.ts
export type OpenAIModel = 'gpt-3.5-turbo' | 'gpt-4' | 'gpt-4-1106-preview' ;
// ...
export interface UmschreibenBody {
  content: string;
  styleValue:
    | ''
    | 'kinderleicht'
    | 'alltagsverständlich'
    | 'gemaessigtFachlich'
    | 'fachspezifisch'
    | 'expertenLevel';
  varianceValue:
    | ''
    | 'Der Text soll sehr nah am Original bleiben und sich kaum verändern!'
    | 'Der Text soll nah am Original bleiben!'
    | ' '
    | 'Der Text darf sich leicht vom Original entfernen!'
    | 'Der Text darf sich sehr stark vom Original entfernen!'
  model: OpenAIModel;
  apiKey?: string | undefined | null;
}